Fixes #146: Avoid using exceptions for flow control in JsonObject implementation
Exceptions by default will fill stacktraces which is very expensive
particularly when performed in a tight loop or in places with deep
stacks (such as on webservers). Replace there use within the variants
of getString, getInt and getBoolean which apply a default value. The
two exceptions which could previously be thrown were null pointers and
class cast which are now both handled using an instanceof check

Bug: https://github.com/eclipse-ee4j/jsonp/issues/146
Signed-off-by: Jason Hinch <jhinch@atlassian.com>