#211: Handle Collections and Maps containing JsonValues
JsonArrayBuilder and JsonObjectBuilder add methods handle
JsonValues and JsonArrayBuilder and JsonObjectBuilder instances,
but creating builders using a Collection or Map that contains
these values fails, because MapUtil does not handle them and
throws an IllegalArgumentException saying that the types are
not supported.

This commit includes tests that create builders with such
values and changes to MapUtil to handle them. (It also
refactors MapUtil to avoid unnecessary excessive nesting.)

Signed-off-by: Todd O'Bryan <toddobryan@gmail.com>
Also-by: Lukas Jungmann <lukas.jungmann@oracle.com>